+++ behavior_machine/library/__init__.py
@@ -2,3 +2,4 @@ from .print_state import PrintState
 from .sequential_state import SequentialState
 from .common_state import IdleState, WaitState
 from .parallel_state import ParallelState
+from .common_state import SetFlowState, SetFlowFromBoardState, SaveFlowState